refactor: Update LOVAKTIVITETER page header font to use Oswald
This commit updates the LOVAKTIVITETER page header font to use the Oswald font. The code changes include importing the Oswald font from the 'next/font/google' package and applying it to the HeaderText component using the styled function from '@mui/system'. This enhancement improves the visual design and typography of the LOVAKTIVITETER page.