refactor: Update package.json dependencies
This commit updates the dependencies in the package.json file. The "@fortawesome/react-fontawesome" dependency has been added with version "^0.2.2" and the "@mui/icons-material" dependency has been updated to version "^5.16.7". These changes ensure that the project is using the latest versions of these libraries and may introduce new features or bug fixes.